added folder type: moved interface and enam
