+++ package.json
@@ -20,6 +20,7 @@
     "axios": "^0.26.1",
     "gh-pages": "^3.2.3",
     "material-ui": "^0.20.2",
+    "material-ui-icons": "^1.0.0-beta.36",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "react-router": "^6.3.0",
