+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Router } from './components/Router';
-import { Header } from './components/Header/components';
+import { Header } from './components/Header';
 import { ThemeProvider } from '@mui/material';
 import { theme } from './theme/theme';
