+++ src/components/Header/constants/socials.ts
@@ -1,20 +1,5 @@
-import {
-  Instagram,
-  LinkedIn,
-  GitHub,
-  SvgIconComponent,
-} from '@mui/icons-material';
-
-export const enum SocialsUrl {
-  INSTAGRAM = 'https://instagram.com/nasa',
-  LINLEDIN = 'https://www.linkedin.com/company/softserve/',
-  GITHUB = 'https://github.com/kerrimov/the-ssspace',
-}
-
-interface Socials {
-  icon: SvgIconComponent;
-  url: SocialsUrl;
-}
+import { Instagram, LinkedIn, GitHub } from '@mui/icons-material';
+import { SocialsUrl, Socials } from '../types/socials.type';
 
 export const socials: Socials[] = [
   { icon: Instagram, url: SocialsUrl.INSTAGRAM },
