+++ src/components/Router/Router.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import './Router.scss';
 import { useRoutes } from 'react-router-dom';
-import { routesConfig } from './routes';
+import { routesConfig } from './roures';
 
 export const Router = () => {
   const routes: React.ReactElement | null = useRoutes(routesConfig);
